Enemy (#10)
* Init Tail

* Fix Racoon attack state

+++ 05-SceneManager/PlayScene.cpp
@@ -130,7 +130,7 @@ void CPlayScene::_ParseSection_OBJECTS(string line)
 	}
 
 	case OBJECT_TYPE_QUESTION_BRICK: {
-		float type = (float)atof(tokens[3].c_str());
+		int type = (int)atof(tokens[3].c_str());
 		obj = new CQuestionBrick(x, y, type);
 		break;
 	}
